Add option for scheduling to leave timeslots alone.
The other options either remake all the timeslots or delete all
the event timeslots. This is not desirable behavior when you have
modified the timeslots or added your own. So this allows you to
edit the start/end time of the event without messing up the
timeslots.

Also, help text is cool.