Initial foray into JavaScript scoring frontend
This is the beautiful beginnings of a javascript frontend to the scoring
system. This is built with one big feature in mind: offline
compatibility. The notion is, if there is a "scoring computer", once it
has loaded the event data, it should be able to track all scores and
produce a scoring slideshow without any internet connection.

To manage this, we build a Backbone.js "score center" which employs the
backbone.dualStorage library. This library wraps Backbone.sync
somewhat-intelligently to automatically, and transparently, fall back to
localStorage. With some fine tuning this seems promising to provide a
complete offline solution.

Many features remain. These include:
-- Changing properties of the events themselves (if it is a trial,
whether to withhold scores, etc.)
-- Producing the scoring slideshow
-- Handling collisions if two people edit the same event simultaneously
-- Improving user feedback when they input bad values.
-- Client-side validation

Nonetheless, this is a powerful proof-of-concept that will be the
beginning of a beautifully modern replacement to the simple form.