Add capability to have static pages with no subdomain
This requires a bit of logic moving -- since the school lookup process
will obviously fail if there is no school to look up, that logic should
be contingent upon finding the school.

This separates the logic, and changes the HomeController to do the right
thing whether or not the school is specified when accessing the
homepage.