Add an event mass-download feature
This was complicated slightly because users can manually modify
timeslots, which will cause new timeslots to be created when the events
are imported again. This creates a new concept of 'custom' value for
Teams per Timeslot and Number of Timeslots that is triggered when the
user deviates from the default timeslot collection of evenly-spaced
timeslots of equal capacity.