Change literally everything
"Move fast and break things."
  -- My Doppleganger

I don't know who reads these commit messages anyway, but here's a short
description of a gigantic commit.

The end goal is to use the asset pipeline. In order to get there, I
needed/wanted to refactor a bunch of stuff to clean up some serious
debt. And I needed to upgrade Rails and sass-rails so that the pipeline
works as expected.

So yeah, pretty much everything is different here and I hope that things
don't break too badly.