Improve #set_current implementation to fix already-current bug
Due to how ActiveRecord caches attributes, the update_attribute call
would not execute unless the record was reloaded. This meant that it
would actually unset the tournament as current.

I added two regression specs and improved the implementation of this
method to return early if there is nothing to be done and also to do the
setting of :is_current in a transaction.